\name{WestGerman}
\alias{WestGerman}
\docType{data}

\title{
Quarterly, West German Investment, Income, and Consumption: 1960Q1-1982Q4
}

\description{Quarterly, seasonally adjusted, West German fixed investment, disposable income, consumption expenditures in billions of DM, 
1960Q1-1982Q4.
}

\usage{data(WestGerman)}
\format{
  A data frame with 92 observations on the following 3 variables.
  \describe{
    \item{\code{invest}}{a numeric vector denotes the investment in billions of DM}
    \item{\code{income}}{a numeric vector denotes the income in billions of DM}
    \item{\code{cons}}{a numeric vector denotes the consumption expenditures in billions of DM}
  }
}

\source{
Deutsche Bundesbank; \url{http://www.jmulti.de/data_imtsa.html}
}

\references{
Lutkepohl, H. (2005). "New introduction to multiple time series analysis". Springer-Verlag, New York.
}

\keyword{datasets}

