\name{portes-package}
\alias{portes-package}
\docType{package}

\title{
Portmanteau Tests for Univariate and Multivariate Time Series Models 
}

\description{
This package contains a set of portmanteau diagnostic checks for univariate and multivariate time series
based on the asymptotic approximation distributions and the Monte-Carlo significance test.
More details about the portmanteau test statistics are available online from the vignette of this package.
This package can be also used to simulate univariate and multivariate data 
from seasonal/nonseasonal \code{ARIMA}/ \code{VARIMA} models with
innovations from finite or infinite variances from stable distributions. 
The simulated data may have deterministic terms, constant drifts and time trends, with non-zero means.
}

\details{
\tabular{ll}{
Package: \tab portes\cr
Type: \tab Package\cr
Version: \tab 5.0\cr
Date: \tab 2020-12-12\cr
LazyLoad: \tab yes\cr
LazyData: \tab yes\cr
Classification/ACM: \tab G.3, G.4, I.5.1\cr
Classification/MSC: \tab 62M10, 91B84\cr
License: \tab GPL (>= 2)\cr
}
}

\author{
Esam Mahdi and A. Ian McLeod.

Maintainer: Esam Mahdi <emahdi@qu.edu.qa>
}

\keyword{ PACKAGE }
\keyword{ ts}




